+++ lib/glimmer/swt/message_box_proxy.rb
@@ -26,6 +26,10 @@ module Glimmer
         end
       end
       
+      def content(&block)
+        Glimmer::DSL::Engine.add_content(self, Glimmer::DSL::SWT::MessageBoxExpression.new, &block)
+      end            
+      
       # TODO refactor the following methods to put in a JavaBean mixin or somethin (perhaps contribute to OSS project too)
       
       def attribute_setter(attribute_name)